backend/requirements.txt
flask
flask-cors
pymongo
bcrypt
python-dotenv
openai